Add chunk aggregation simulation
Chunk aggregation (or frame aggregation if you're simulating sending frames) aggregates N chunks and sends them as a single chunk of bits.

While from the sending perspective this is exactly the same as sending a chunk N times bigger, from the receiving perspective it is not.
Different chunks can be recover or discarded and account as a loss or received.

To make it clear to control the simulation, send_max parameter was renamed to tx_max, making it functionally similar, but easier to use and understand.